requirements.txt
# Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# WebSocket
python-socketio[client]==5.11.0
eventlet==0.33.3

# AI & ML
google-generativeai==0.3.2
Pillow==10.2.0

# Image Capture
mss==9.0.1
psutil==5.9.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Security
cryptography==41.0.7

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
